Split out termui
